+++ Tests/Unit/Testing/FrameworkTest.php
@@ -5,7 +5,9 @@ declare(strict_types=1);
 namespace OliverKlee\Oelib\Tests\Unit\Testing;
 
 use Nimut\TestingFramework\TestCase\UnitTestCase;
+use OliverKlee\Oelib\System\Typo3Version;
 use org\bovigo\vfs\vfsStream;
+use TYPO3\CMS\Core\Core\Environment;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
@@ -80,7 +82,11 @@ class FrameworkTest extends UnitTestCase
      */
     public function cleanUpWithoutDatabaseDeletesCreatedDummyUploadFolder()
     {
-        $this->subject->setUploadFolderPath(PATH_site . 'typo3temp/tx_oelib_test/');
+        if (Typo3Version::isNotHigherThan(8)) {
+            $this->subject->setUploadFolderPath(PATH_site . 'typo3temp/tx_oelib_test/');
+        } else {
+            $this->subject->setUploadFolderPath(Environment::getPublicPath() . '/typo3temp/tx_oelib_test/');
+        }
         $this->subject->createDummyFile();
 
         self::assertDirectoryExists($this->subject->getUploadFolderPath());
@@ -209,7 +215,11 @@ class FrameworkTest extends UnitTestCase
      */
     public function createDummyFileForNonExistentUploadFolderSetCreatesUploadFolder()
     {
-        $this->subject->setUploadFolderPath(PATH_site . 'typo3temp/tx_oelib_test/');
+        if (Typo3Version::isNotHigherThan(8)) {
+            $this->subject->setUploadFolderPath(PATH_site . 'typo3temp/tx_oelib_test/');
+        } else {
+            $this->subject->setUploadFolderPath(Environment::getPublicPath() . '/typo3temp/tx_oelib_test/');
+        }
         $this->subject->createDummyFile();
 
         self::assertDirectoryExists($this->subject->getUploadFolderPath());
@@ -220,7 +230,11 @@ class FrameworkTest extends UnitTestCase
      */
     public function createDummyFileForNonExistentUploadFolderSetCreatesFileInCreatedUploadFolder()
     {
-        $this->subject->setUploadFolderPath(PATH_site . 'typo3temp/tx_oelib_test/');
+        if (Typo3Version::isNotHigherThan(8)) {
+            $this->subject->setUploadFolderPath(PATH_site . 'typo3temp/tx_oelib_test/');
+        } else {
+            $this->subject->setUploadFolderPath(Environment::getPublicPath() . '/typo3temp/tx_oelib_test/');
+        }
         $dummyFile = $this->subject->createDummyFile();
 
         self::assertFileExists($dummyFile);
@@ -310,7 +324,11 @@ class FrameworkTest extends UnitTestCase
      */
     public function createDummyFolderForNonExistentUploadFolderSetCreatesUploadFolder()
     {
-        $this->subject->setUploadFolderPath(PATH_site . 'typo3temp/tx_oelib_test/');
+        if (Typo3Version::isNotHigherThan(8)) {
+            $this->subject->setUploadFolderPath(PATH_site . 'typo3temp/tx_oelib_test/');
+        } else {
+            $this->subject->setUploadFolderPath(Environment::getPublicPath() . '/typo3temp/tx_oelib_test/');
+        }
         $this->subject->createDummyFolder('test_folder');
 
         self::assertDirectoryExists($this->subject->getUploadFolderPath());
@@ -321,7 +339,11 @@ class FrameworkTest extends UnitTestCase
      */
     public function createDummyFolderForNonExistentUploadFolderSetCreatesFileInCreatedUploadFolder()
     {
-        $this->subject->setUploadFolderPath(PATH_site . 'typo3temp/tx_oelib_test/');
+        if (Typo3Version::isNotHigherThan(8)) {
+            $this->subject->setUploadFolderPath(PATH_site . 'typo3temp/tx_oelib_test/');
+        } else {
+            $this->subject->setUploadFolderPath(Environment::getPublicPath() . '/typo3temp/tx_oelib_test/');
+        }
         $dummyFolder = $this->subject->createDummyFolder('test_folder');
 
         self::assertDirectoryExists($dummyFolder);
@@ -337,7 +359,7 @@ class FrameworkTest extends UnitTestCase
     public function getUploadFolderPathReturnsUploadFolderPathIncludingTablePrefix()
     {
         self::assertRegExp(
-            '/\\/uploads\\/tx_oelib\\/$/',
+            '/\\/typo3temp\\/tx_oelib\\/$/',
             $this->subject->getUploadFolderPath()
         );
     }
@@ -387,7 +409,11 @@ class FrameworkTest extends UnitTestCase
             'The first parameter $absolutePath is not within the calling extension\'s upload directory.'
         );
 
-        $this->subject->getPathRelativeToUploadDirectory(PATH_site);
+        if (Typo3Version::isNotHigherThan(8)) {
+            $this->subject->getPathRelativeToUploadDirectory(PATH_site);
+        } else {
+            $this->subject->getPathRelativeToUploadDirectory(Environment::getPublicPath() . '/');
+        }
     }
 
     /*